# frontend/dist/openapi.yaml
openapi: 3.1.0
info:
  title: Themis API
  version: 0.1.0
  description: |
    Themis network operations API for devices, change requests, approvals, comments,
    golden configs, drift detection, RBAC, AD/LDAP, OAuth/OIDC, and safe deployments.
servers:
  - url: /api
    description: Same-origin Themis API
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Log in with a local or AD/LDAP account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT and user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register a local account
      description: The first registered user becomes admin. Later users default to engineer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: JWT and user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/password:
    put:
      tags: [Auth]
      summary: Change current user's local password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /devices/{id}/drift:
    parameters:
      - $ref: "#/components/parameters/DeviceId"
    get:
      tags: [Drift]
      summary: Get open drift for a device
      responses:
        "200":
          description: Drift or null
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ConfigDrift"
                  - type: "null"
...
  /drift/{id}/accept:
    parameters:
      - $ref: "#/components/parameters/DriftId"
    post:
      tags: [Drift]
      summary: Accept drift and create a change request
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
      responses:
        "200":
          description: Updated drift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigDrift"
  /drift/{id}/dismiss:
    parameters:
      - $ref: "#/components/parameters/DriftId"
    post:
      tags: [Drift]
      summary: Dismiss drift
      responses:
        "200":
          description: Updated drift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigDrift"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DeviceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    DriftId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    UserRole:
      type: string
      enum: [viewer, engineer, admin]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
      required: [id, username, email, role]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: password
      required: [username, password]
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required: [username, email, password]
    ChangePasswordRequest:
      type: object
      properties:
        current_password:
          type: string
          format: password
        new_password:
          type: string
          format: password
      required: [current_password, new_password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
...
    ConfigDrift:
      type: object
      properties:
        id:
          type: string
          format: uuid
        device_id:
          type: string
          format: uuid
        device_name:
          type: string
        golden_config_id:
          type: string
          format: uuid
        current_config:
          type: string
        detected_at:
          type: string
          format: date-time
        last_checked_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [open, accepted, dismissed, auto_resolved]
        resolved_by:
          type: string
          format: uuid
          nullable: true
        resolved_by_username:
          type: string
          nullable: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
        accepted_change_id:
          type: string
          format: uuid
          nullable: true
      required: [id, device_id, device_name, golden_config_id, current_config, detected_at, last_checked_at, status]
...